% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_summary.R
\name{beta_summary}
\alias{beta_summary}
\title{Returns the total bias, null rejection rate, coverage, and effect size mean and standard deviation}
\usage{
beta_summary(run, a = 0.05)
}
\arguments{
\item{run}{a  list object as returned by varied_runs() or reparse_runs()}

\item{a}{The alpha value for calculating the null rejection rate, defaults to 0.05}
}
\value{
A data frame with the following summary statistics listed above
}
\description{
Specifically beta_summary returns:
\itemize{
\item Total bias: the average effect size estimate - the set effect size. Also returns its standard error
\item Null rejection rate: the percentage of p values that were less than the set alpha value. Also returns its standard error
\item Coverage: the percent of effect size estimates that contained the set effect size in its 95\% confidence interval. Also returns its standard error
\item The mean effect size estimate
\item The standard deviation of the effect size estimate
}
}
\examples{
#creates the causal diagram
my_dag = HydeNetwork(~B|A + C|A + D|B*C)

#adds set effect sizes and distributions to the causal diagram
my_dag = setNode(my_dag, A, nodeType = "dbern", prob = 0.5)
my_dag = setNode(my_dag, B, nodeType = "dnorm", mu = paste0(0.2," * A + ",0 - 0.2*0.5), tau = 1)
my_dag = setNode(my_dag, C, nodeType = "dbern", prob = paste0("ilogit(",0.8," * A + ",set_p(0.9, 0.8*0.5),")"))
my_dag = setNode(my_dag, D, nodeType = "dbern", prob = paste0("ilogit(",0.05," * B + ", 1.2," * C + ",set_p(0.25, 1.2 * 0.9 + 0.05 * 0 ),")"))

#ScenarioMatrix object
ideal_run = varied_runs(runs = 100, dag = my_dag, exposure = "B", outcome = "C", covariates = "A", n = 1000)

#Gives a dataframe of values for quantifying the bias and its effects on the reject null rate
beta_summary(ideal_run)

#Creates a ridgeline plot to help with bias and coverage visualization
ci_ridges(ideal_run)
}
\seealso{
\code{\link[=varied_runs]{varied_runs()}}, \code{\link[=ci_ridges]{ci_ridges()}}, \code{\link[=reparse_runs]{reparse_runs()}}
}
