% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydenet_bindPosterior.R
\name{bindSim}
\alias{bindSim}
\alias{bindPosterior}
\title{Bind Output From coda Samples}
\usage{
bindSim(hydeSim, relabel_factor = TRUE)

bindPosterior(hydeSim, relabel_factor = TRUE)
}
\arguments{
\item{hydeSim}{An object of class \code{HydeSim}}

\item{relabel_factor}{Logical.  If \code{TRUE}, factors that had been
converted to integers for the JAGS code can be relabelled as factors
for additional analysis in R.}
}
\description{
After determining the simulated distributions are satisfactory,
it can be advantageous to bind the simulated distributions together in
order to aggregate values and perform other manipulations and analyses.
}
\details{
For the purposes of this function, it is assumed that if the
simulated distributions are satisfactory, the multiple chains in a run
can be bound together.  Subsequently, the multiple runs are bound
together.  Lastly, the factors are relabeled, if requested.
}
\examples{
#creates the causal diagram
my_dag = HydeNetwork(~B|A + C|A + D|B*C)

#adds set effect sizes and distributions to the causal diagram
my_dag = setNode(my_dag, A, nodeType = "dbern", prob = 0.5)
my_dag = setNode(my_dag, B, nodeType = "dnorm", mu = paste0(0.2," * A + ",0 - 0.2*0.5), tau = 1)
my_dag = setNode(my_dag, C, nodeType = "dbern", prob = paste0("ilogit(",0.8," * A + ",set_p(0.9, 0.8*0.5),")"))
my_dag = setNode(my_dag, D, nodeType = "dbern", prob = paste0("ilogit(",0.05," * B + ", 1.2," * C + ",set_p(0.25, 1.2 * 0.9 + 0.05 * 0 ),")"))

#ScenarioMatrix object
ideal_run = varied_runs(runs = 100, dag = my_dag, exposure = "B", outcome = "C", covariates = "A", n = 1000)

#Gives a dataframe of values for quantifying the bias and its effects on the reject null rate
beta_summary(ideal_run)

#Creates a ridgeline plot to help with bias and coverage visualization
ci_ridges(ideal_run)
}
\author{
Jarrod Dalton and Benjamin Nutter
}
