% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{Create simulated data from a DAG object}
\usage{
create_data(dag, n, positivity = F, ...)
}
\arguments{
\item{dag}{Hydenet directed acyclic graph object (DAG object) with set distributions for each node.}

\item{n}{Integer, the number of samples in the simulated data frame}

\item{positivity}{logical. If set to TRUE, checks for positivity violations among binary columns. If violated, it changes the value in the first row of the column to comply with positivity. Defaults to FALSE.}

\item{...}{if variables are written into the DAG object, they must all be numerically set here.
This allows the user to quickly change DAG values when generating simulated data}
}
\value{
This function returns a simulated data frame. Each column in the data frame corresponds to a node
in the DAG object. The data frame will have n rows. Nodes with continuous distributions will have a
double class, and nodes with binary distributions will have an integer class.
}
\description{
Creates a single data frame of a set sample size from a DAG object with set distributions
}
\examples{
examples/create_data_example.R
}
