% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misdiagnosis.R
\name{misdiagnosis}
\alias{misdiagnosis}
\title{Changes a data frame's values to represent a misdiagnosis rate}
\usage{
misdiagnosis(df, variable, under_rate = 0, over_rate = 0)
}
\arguments{
\item{df}{data frame with at least one binary column set as an integer}

\item{variable}{Character value. The column in the data frame that the misdiagnosis rate will be set to. Must be integer class and only contain 0s or 1s.}

\item{under_rate}{The percentage of 0s in the column the user wishes to change to 1s}

\item{over_rate}{The percentage of 1s in the column the user wishes to change to 0s}
}
\value{
A data frame with the values of the specifed column changed as set by the user
}
\description{
Changes values of a specified binary column in a data frame to simulate information bias and or add noise. If under_rate is set it will change a percentage of 1s to 0s and vice versa if over_rate is set. Both under_rate and over_rate can be set.
}
\examples{
examples/misdiagnosis_example.R
}
\seealso{
\code{\link[=varied_runs]{varied_runs()}}
}
