% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparse_runs.R
\name{reparse_runs}
\alias{reparse_runs}
\title{Combines outputs from varied_runs() together to create a new output that contains different analyses to compare.}
\usage{
reparse_runs(
  run_list,
  method = NULL,
  list_names = as.character(1:length(run_list))
)
}
\arguments{
\item{run_list}{Several different outputs from varied_runs() in a list format that the user wishes to recombine. The outputs from varied_runs() in the list must all have the same number of runs (the same number of datasets must be simulated).}

\item{method}{Character value. Defaults to Null, meaning every method in each varied_runs() output will be included.}

\item{list_names}{Character vector. Must be equal to the number of elements in run_list. Defaults to numeric naming with the first element of run_list being named "1", the second "2" and so on.}
}
\value{
A list of 11 elements with each element bulleted below in order
\itemize{
\item odds_ratio, NA if not applicable to the method used
\item calculated_ate, the calculated effect size estimate between the set exposure and outcome
\item lower_int, the lower bound of the 95\% confidence interval for the calculated_ate
\item upper_int, the upper bound of the 95\% confidence interval for the calculated_ate
\item p_values, the p value for the calculated_ate
\item exp_prevalence, the exposure prevalence if the exposure variable is binary, returns NAs if not
\item out_prevalence, the outcome prevalence if the outcome variable is binary, returns NAs if not
\item sample_population, the number of samples in the dataset. If the analysis requires stratifying or throwing data samples, sample_population will change to reflect that.
\item set_ate, the effect size between the exposure and outcome that is directly set in the DAG object. Acts as the true effect size.
\item over_r, the overdiagnosis rate that was set. defaults to 0
\item under_r, the overdiagnosis rate that was set. defaults to 0
}

each list element will contain x number of columns and y number of rows, where x is the number of different methods tested, and y is the number of simulated data sets created.
}
\description{
The final output will be a list of 11 elements with each element corresponding to a summary statistic, but the columns in each element will now reflect the different circumstances of each scenario set from varied_runs().
}
\examples{
examples/varied_methods_example.R
}
\seealso{
\code{\link[=apply_methods]{apply_methods()}},\code{\link[=ci_ridges]{ci_ridges()}},\code{\link[=beta_summary]{beta_summary()}}, \code{\link[=varied_runs]{varied_runs()}}
}
