% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_ratio.R
\name{risk_ratio}
\alias{risk_ratio}
\title{Analyzes a data frame to give an risk ratio}
\usage{
risk_ratio(exposure, outcome, covariates = NULL, df)
}
\arguments{
\item{exposure}{Character value. The name of a column in a data frame that represents exposure in an}

\item{outcome}{Character value. The name of a \strong{binary} column in a data frame that represents the outcome.}

\item{covariates}{Character value/vector. The name(s) of the columns that the user wants adjust on.}

\item{df}{The data frame for analysis}
}
\value{
A six column data frame with the following summary statistics:
\itemize{
\item risk ratio
\item beta (log odds)
\item lower 95\% confidence interval of beta coefficient
\item upper 95\% confidence interval of beta coefficient
\item p value
\item number of samples in the dataset
}
}
\description{
returns a relative risk value via the delta method offered in the logisticRR package. The equation is created with the following parameters arranged like so: outcome ~ exposure + covariates.
}
\examples{
examples/effect_estimate_functions.R
}
\seealso{
\code{\link[=varied_runs]{varied_runs()}}, \code{\link[=apply_methods]{apply_methods()}}
}
