% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2d_dag.R
\docType{data}
\name{t2d_dag}
\alias{t2d_dag}
\title{A hydenet object with the set distributions shown in the paper use case}
\format{
A HydeNetwork with 7 nodes and set node distributions
\describe{
\item{BMI}{Gaussian distribution node reflecting body mass index}
\item{WC}{Gaussian distribution node reflecting waist circumference}
\item{SBP}{Gaussian distribution node reflecting systolic blood pressure}
\item{DM}{binary distribution reflecting type 2 diabetes prevalence}
\item{PAD}{binary distribution reflecting peripherial artery disease prevalence}
\item{CHD}{binary distribution reflecting coranary heart disease prevalence}
\item{rs7903146}{binary distribution reflecting the prevalence of the single nucleotide polymorphism rs7903146}
}
}
\usage{
t2d_dag
}
\description{
A DAG object that represents a predictive model of type 2 diabetes development and development of conditions related to type 2 diabetes.
}
\keyword{datasets}
