% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vignette_documentation.R
\docType{data}
\name{UseCase1Results.rdata}
\alias{UseCase1Results.rdata}
\title{Simulation Results for Use Case 1}
\format{
outcome Matrices and DAG
}
\usage{
data(UseCase1Results)
}
\description{
An R data file with the results reported in Use Case 1 in McBias's paper
}
\keyword{data}
