% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData_documentaion.R
\docType{data}
\name{UseCase2Results.RData}
\alias{UseCase2Results.RData}
\title{Use Case 2 Output data, Demonstrates how different adjustment sets affect bias.}
\format{
Two list outputs from reparse_runs() with 11 elements where each element is a 2 dimensional matrix of 500x3 where each column represents each of the 3 adjustment sets and each row represents the calculation for each of the 500 iterations
\describe{
\item{ratio}{Odds ratio. Reads as NA since the ouctome is continuous}
\item{calculated_ate}{The effect size calculated from the simulated datasets with the corresponding adjustment set}
\item{lower_int}{The lower 95\% confidence interval of the calculated effect size}
\item{upper_int}{The upper 95\% confidence interval of the calculated effect size}
\item{p_values}{The p value of the calculated effect size}
\item{exp_prevalence}{The prevalence of the exposure (rs7903146T)}
\item{out_prevalence}{The prevalence of the outcome, NA since outcome is continuous}
\item{sample_population}{sample size of the simulated dataset.
The "naive, UKB selection bias" sample sizes vary slightly since it reports the number of samples remaining after "In_UKB" is stratified}
\item{set_ate}{the effect size the BN}
\item{over_r}{percentage of 0s changed to 1s in a selected binary vairable. defaults to 0}
\item{under_r}{percentage of 1s changed to 0s in a selected binary vairable. defaults to 0}
}
}
\usage{
load("UseCase2Results.RData")
}
\description{
The output matrix data used for constructing the use case figures.
}
\details{
tot_bmi_run columns represent the following:
naive is BMI ~ Rs303752
naive, UKB selection bias is BMI ~ Rs303752 where In_UKB is stratified

tot_edu_run columns represent the following:
naive is Educational_age ~ Rs303752
naive, UKB selection bias is Educational_age ~ Rs303752 where In_UKB is stratified
}
\keyword{data}
