% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vignette_documentation.R
\docType{data}
\name{vignette_data.rdata}
\alias{vignette_data.rdata}
\title{Simulations for vignette}
\format{
outcome Matrices and DAG
}
\usage{
data(vignette_data)
}
\description{
An R data file for running the vignette
}
\keyword{data}
