% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vignette_documentation.R
\docType{data}
\name{vignette_data2.RData}
\alias{vignette_data2.RData}
\title{More Simulations for different methods vignette}
\format{
outcome Matrices and DAG
}
\usage{
data(vignette_data2)
}
\description{
An R data file for running the vignette
}
\keyword{data}
