% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_methods.R
\name{apply_methods}
\alias{apply_methods}
\title{Analyzes a data frame with various methods and reports the summary statistics}
\usage{
apply_methods(
  exposure,
  outcome,
  covariates = NULL,
  sb = NULL,
  df,
  ratio = 1,
  match_methods = NULL
)
}
\arguments{
\item{exposure}{Character value. The column name in a data frame that represents exposure in an}

\item{outcome}{Character value. The column name in a data frame that represents the outcome.}

\item{covariates}{Character value/vector. The name(s) of the columns that the user wants adjust on. Defaults to NULL meaning no covariates are adjusted on.}

\item{sb}{Character value/vector. The name(s) of \strong{binary} column(s). Stratifies the data frame to include only the rows with 1s in every listed column.}

\item{df}{The data frame for analysis}

\item{ratio}{Integer. number of controls that should be matched to each case. Defaults to 1}

\item{match_methods}{Character value/vector. The method(s) in which distance between a matched case and control is measured as specified by MatchIt. Defaults to NULL which does no matching analysis.}
}
\value{
A six column data frame of n rows where n is the number of different methods run. The columns represent the following summary statistics:
\itemize{
\item odds ratio
\item beta (log odds)
\item lower 95\% confidence interval of beta coefficient
\item upper 95\% confidence interval of beta coefficient
\item p value
\item number of samples in the dataset
}
}
\description{
runs the appropriate methods based on the outcome's distribution. Also can run matching methods if set by the user. Returns the summary statistics of each analysis to help with method comparison.
}
\examples{
examples/apply_methods_to_matrix.R
}
\seealso{
\code{\link[=varied_runs]{varied_runs()}}, \code{\link[=lm_beta]{lm_beta()}}, \code{\link[=odds_ratio]{odds_ratio()}}, \code{\link[=get_ps]{get_ps()}}, \code{\link[=matchit_matching]{matchit_matching()}}, \code{\link[=ps_weight]{ps_weight()}}, \code{\link[=risk_ratio]{risk_ratio()}}
}
