% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ps.R
\name{get_ps}
\alias{get_ps}
\title{Interfaces with the MatchIt package to create a matched dataset as specifed by the user}
\usage{
get_ps(exposure, covariates, df)
}
\arguments{
\item{exposure}{Character value. The name of the column in a data frame that represents the exposure}

\item{covariates}{Character value/vector. The name(s) of the columns that the user wants to adjust on.}

\item{df}{The data frame set for calculating propensity score values}
}
\value{
The original data frame with an additional propensity score column and an additional propensity score weight column
}
\description{
Uses MatchIt package as a way of adjusting for covariates. Matches cases to controls from the set "exposure" column
Note that users are limited to the greedy "nearest" method, though they may enter in different distance calculation methods
}
\examples{
examples/propensity_score_df.R
}
\seealso{
\code{\link[=varied_runs]{varied_runs()}}, \code{\link[=applied_methods]{applied_methods()}}
}
