% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchit_matching.R
\name{matchit_matching}
\alias{matchit_matching}
\title{Interfaces with the MatchIt package to create a matched dataset as specifed by the user}
\usage{
matchit_matching(exposure, covariates, df, d = "logit", ratio = 1)
}
\arguments{
\item{exposure}{Character value. The name of a binary column in a data frame that represents the cases and controls that need to be matched}

\item{covariates}{Character value/vector. The name(s) of the columns that the user wants to match the cases and controls on.}

\item{df}{The data frame set for matching}

\item{d}{the method in which distance between a matched case and control is measured as specified by MatchIt. Defaults to "logit"}

\item{ratio}{Integer. number of controls that should be matched to each case. Defaults to 1}
}
\value{
A data frame where each case from the exposure column is matched to a set number of controls based on the named covariates.
}
\description{
Uses MatchIt package as a way of adjusting for covariates. Matches cases to controls from the set "exposure" column
Note that users are limited to the greedy "nearest" method, though they may enter in different distance calculation methods
}
\examples{
examples/MatchIt_example.R
}
\seealso{
\code{\link[=varied_runs]{varied_runs()}}
}
