% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_weight.R
\name{ps_weight}
\alias{ps_weight}
\title{Analyzes a data frame to give an odds ratio}
\usage{
ps_weight(exposure, outcome, covariates, df, weights)
}
\arguments{
\item{exposure}{Character value. The name of a column in a data frame that represents exposure in an}

\item{outcome}{Character value. The name of a column in a data frame that represents the outcome.}

\item{covariates}{Character value/vector. The name(s) of the columns that the user wants adjust on.}

\item{df}{The data frame for analysis. must have a column that includes weights. Users may generate a weight column with get_ps()}

\item{weights}{Character value. The name of a column in a data frame that represents the weights the user wishes to use.}
}
\value{
A six column data frame with the following summary statistics:
\itemize{
\item odds ratio (reads NA if outcome is Gaussian)
\item beta coeffecient
\item lower 95\% confidence interval of beta coefficient
\item upper 95\% confidence interval of beta coefficient
\item p value
\item number of samples in the dataset
}
}
\description{
runs a linear or logistic regression depending on if the outcome is Gaussian or binary. Adjusts for covariates with given weights
The regression is created with the following parameters ordered like so: outcome ~ exposure + covariates.
}
\examples{
#creates the causal diagram
confounder_dag = HydeNetwork(~B|A + C|A + D|B*C)

#adds set effect sizes and distributions to the causal diagram
confounder_dag = setNode(confounder_dag, A, nodeType = "dbern", prob = 0.5)
confounder_dag = setNode(confounder_dag, B, nodeType = "dnorm", mu = paste0(1," * A + ",0 - 1*0.5), tau = 1)
confounder_dag = setNode(confounder_dag, C, nodeType = "dbern", prob = paste0("ilogit(",0.8," * A + ",set_p(0.1, 0.8*0.5),")"))
confounder_dag = setNode(confounder_dag, D, nodeType = "dbern", prob = paste0("ilogit(",0.05," * B + ", 1.2," * C + ",set_p(0.25, 1.2 * 0.1 + 0.05 * 0 ),")"))

#create data creates a single dataset.
#This lets users see if the simulated datasets behave as expected before committing to a large simulation with many iterations

sim_data = create_data(confounder_dag, 10000)

#outcome must be binary
or_result = odds_ratio(exposure = "A", outcome = "D", covariates = c("B","C"), df = sim_data)


#outcome must be continuous
lm_results = lm_beta(exposure = "C", outcome = "B", covariates = "A", df = sim_data)

#outcome may be either
ps_df = get_ps(exposure = "C", covariates = "A", df = sim_data)
ps_results = ps_weight(exposure = "C", outcome = "B", covariates = "A", df = ps_df, weights = "weights")
}
\seealso{
\code{\link[=varied_runs]{varied_runs()}},\code{\link[=apply_methods]{apply_methods()}}, \code{\link[=get_ps]{get_ps()}}
}
