% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_weight.R
\name{ps_weight}
\alias{ps_weight}
\title{Analyzes a data frame to give an odds ratio}
\usage{
ps_weight(exposure, outcome, covariates, df, weights)
}
\arguments{
\item{exposure}{Character value. The name of a column in a data frame that represents exposure in an}

\item{outcome}{Character value. The name of a column in a data frame that represents the outcome.}

\item{covariates}{Character value/vector. The name(s) of the columns that the user wants adjust on.}

\item{df}{The data frame for analysis. must have a column that includes weights. Users may generate a weight column with get_ps()}

\item{weights}{Character value. The name of a column in a data frame that represents the weights the user wishes to use.}
}
\value{
A six column data frame with the following summary statistics:
\itemize{
\item odds ratio (reads NA if outcome is Gaussian)
\item beta coeffecient
\item lower 95\% confidence interval of beta coefficient
\item upper 95\% confidence interval of beta coefficient
\item p value
\item number of samples in the dataset
}
}
\description{
runs a linear or logistic regression depending on if the outcome is Gaussian or binary. Adjusts for covariates with given weights
The regression is created with the following parameters ordered like so: outcome ~ exposure + covariates.
}
\examples{
examples/effect_estimate_functions.R
}
\seealso{
\code{\link[=varied_runs]{varied_runs()}},\code{\link[=apply_methods]{apply_methods()}}, \code{\link[=get_ps]{get_ps()}}
}
