% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vignette_documentation.R
\docType{data}
\name{UseCase2Results.rdata}
\alias{UseCase2Results.rdata}
\title{Simulation Results for Use Case 2}
\format{
outcome Matrices and DAG
}
\usage{
data(UseCase2Results)
}
\description{
An R data file with the results reported in Use Case 2 in McBias's paper
}
\keyword{data}
